requirements.txt
emoji
terminal-animation
colorama